pyproject.toml
[project]
name = "notificationservice"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework>=3.16.1",
]
